requirements.txt
# Global Regime Rotator - Dependencies
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
yfinance>=0.2.36
requests>=2.31.0
pyarrow>=14.0.0
pytest>=7.4.0
